Altered GroupUserDetails and InterestPoints models. Created routes to support group map grabbing of users and interest points as well as locating preferences.

+++ models/InterestPoints.js
@@ -8,8 +8,8 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false,
         },
         description: DataTypes.STRING,
-        longitude: DataTypes.FLOAT,
-        latitude: DataTypes.FLOAT,
+        longitude: DataTypes.DOUBLE,
+        latitude: DataTypes.DOUBLE,
         photo_url: DataTypes.STRING,
         color: {
             type: DataTypes.STRING,
